+++ packages/home-trmnl-core/src/repositories/device-repository.js
@@ -1,6 +1,3 @@
-import {
-	randomUUID
-} from 'node:crypto';
 import debug from 'debug';
 
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - -
@@ -9,65 +6,20 @@ const log = debug('home-trmnl:core:device-repository');
 
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
-function createScreenTable (definitions)
-{
-	const screens = [];
-
-	for (const { type, duration, layout, panels, image } of definitions)
-	{
-		const composed = type === 'composed', id = randomUUID();
-
-		screens.push({
-			id,
-			composed,
-			panels,
-			duration,
-			layout,
-			image
-		});
-
-		log('Loaded %s screen into data table.', type);
-	}
-
-	return screens;
-}
-
-function createDeviceTable (definitions)
-{
-	const devices = {};
-
-	for (const { id, address, key, screens, autoUpdate, bitDepth } of definitions)
-	{
-		devices[address] = {
-			id,
-			address,
-			key,
-			bitDepth,
-			autoUpdate,
-			screen   : -1,
-			model    : null,
-			firmware : null,
-			battery  : 0,
-			rssi     : -100,
-			error    : false,
-			screens  : createScreenTable(screens)
-		};
-
-		log('Loaded device `%s`, with address `%s`, into data table.', id, address);
-	}
-
-	return devices;
-}
-
-// - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-
 export class DeviceRepository
 {
 	#devices = {};
 
-	constructor (devices)
+	async addDevice (device)
 	{
-		this.#devices = createDeviceTable(devices);
+		const existing = this.#devices[device.address];
+
+		if (existing != null)
+		{
+			throw new Error(`Cannot add device '${device.id}', it has the same MAC address as device '${existing.id}'.`);
+		}
+
+		this.#devices[device.address] = device;
 	}
 
 	async getDevice (address)
@@ -81,25 +33,23 @@ export class DeviceRepository
 
 		const {
 			id,
-			model,
 			firmware,
 			battery,
 			rssi,
-			bitDepth,
 			autoUpdate,
-			error
+			error,
+			model
 		} = device;
 
 		return {
 			id,
 			address,
-			model,
 			firmware,
 			battery,
 			rssi,
-			bitDepth,
 			autoUpdate,
-			error
+			error,
+			model
 		};
 	}
 
@@ -113,30 +63,7 @@ export class DeviceRepository
 		return this.devices[address]?.screens || [];
 	}
 
-	async updateDeviceToNextScreen (address)
-	{
-		const device = this.#devices[address];
-
-		if (!device)
-		{
-			return null;
-		}
-
-		let next = device.screen + 1;
-
-		if (next >= device.screens.length)
-		{
-			next = 0;
-		}
-
-		// Record.
-		device.screen = next;
-
-		return device.screens[next];
-	}
-
 	async updateDevice (address, {
-		model,
 		firmware,
 		battery,
 		rssi,
@@ -148,13 +75,6 @@ export class DeviceRepository
 		if (!device)
 		{
 			return;
-		}
-
-		if (model != null)
-		{
-			device.model = model;
-
-			log('Updating device, with address `%s`, to have model `%s`.', address, model);
 		}
 
 		if (firmware != null)
@@ -184,5 +104,26 @@ export class DeviceRepository
 
 			log('Updating device, with address `%s`, to have error status of `%s`.', address, error);
 		}
+	}
+
+	async updateDeviceToNextScreen (address)
+	{
+		const device = this.#devices[address];
+
+		if (!device)
+		{
+			return null;
+		}
+
+		let next = device.screen + 1;
+
+		if (next >= device.screens.length)
+		{
+			next = 0;
+		}
+
+		device.screen = next;
+
+		return device.screens[next];
 	}
 }