+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.5.0 - Unreleased
+
+## Added
+
+- Added the `model` device configuration that will result in better rendering for officially supported TRMNL devices, as documented in the [TRMNL Styleguide](https://trmnl.com/framework/docs/).
+
+## Removed
+
+- Removed the `bitDepth` device configuration in favour of the new `model` property.
+
 ## 0.4.3 - 2026-02-15
 
 The first documented release.